Create Check Mate .cpp
You are given a standard 8×8 chessboard which has exactly 3 pieces placed on it - 2 black rooks and 1 white king. The rows are numbered 1 to 8 from bottom to top, and the columns are numbered 1 to 8 from left to right. The cell at the intersection of the i-th column and j-th row is denoted (i,j) (like the coordinate system on the xy-plane).

Is it possible for the white king to be checkmated by the rooks in exactly one move? It is guaranteed that the initial position of pieces is such that the white king is not under check.

Note that in the given situation, black moves first.

Assume that all the chess rules are to be followed:

Each square can contain at most one piece.
Rooks can only travel in straight lines along the row or column they are placed at, and can't jump over other pieces.
The King can move to any of its 8 adjacent squares.
A piece can't capture a piece of the same color as itself.
After a black makes a move, the white king can capture a rook if it's present on any of the 8 adjacent squares and not protected by the other rook.
The king is said to be in checkmate if the king is currently under check and there exists no legal move with the king to escape from the check.
Stalemate is considered as a draw for both sides.