Create Rock Paper Scissors.cpp
There are N players standing in a line, indexed 1 to N from left to right. They all play a game of Rock, Paper, Scissors. Each player has already decided which move they want to play. You are given this information as a string S of length N, i.e,

Si is equal to R if player i will play Rock.
Si is equal to P if player i will play Paper.
Si is equal to S if player i will play Scissors.
Let W(i,j) denote the move played by the winner if players i,i+1,…,j compete in order from left to right. That is,

First, players i and i+1 play a game
The winner of this game plays against player i+2
The winner of the second game plays against player i+3
⋮
The winner of the first j−i−1 games plays against player j, and the move played by the winner of this game is declared to be W(i,j).
If i=j, then player i is considered to be the winner and W(i,i)=Si.

Your task is to find the value of W(i,N) for all i from 1 to N.